flaskapp/app/requirements.txt
# Core dependencies for structured data analysis
requests>=2.28.0
extruct>=0.12.0
w3lib>=2.1.0
lxml>=4.9.0

# Flask API server dependencies
Flask>=2.3.0
Flask-CORS>=4.0.0

# Enhanced analysis dependencies
jsonschema>=4.17.0

# Environment and configuration
python-dotenv>=1.0.0

# Type hints and validation
typing-extensions>=4.0.0